Extract test fixture for reuse
